#104 Created separate node.js server to run scripts on
